refactor: move answer GET & POST to answers.py & include username in route
